Allow passing env var GUNICORN_WORKERS
This allows us to configure via an env var the number of gunicorn
workers, making it consistent with the other stuff we configure via env
vars. Switch from the not easy to reason about and debug
multiprocessing.cpu_count() to 1, which is easier. We 'll override with
specific values in production

Bug: T331505
Change-Id: I62e2472e90b69005734fbfdc57af22e915f07af5