Fix the tokenization in NLLB Wikipedia model
The source language at the end of tokenized input was causing slightly
incorrect translations.

For example. "water" was getting translated to igbo as "mmiri na-eme ka ọ dị"
instead of "mmiri". "water" was getting translated to Assamese as
"পানী ব্যৱহাৰ" instead of "পানী"

This was also observed when comparing the output of the above words with the
unoptimized NLLB 200 model(currently hosted in AWS)

Change-Id: I598c214303dc2beb05be1c0d12c139c3d75dd178