performance tuning: Set CT2_INTRA_THREADS to 0
Leave CT2_INTRA_THREADS as 0 so that CTranslate2 use a default value.
Found that this is the best setting for performance than we setting any
value to it.

Also set CT2_INTER_THREADS to number of CPUs as that is the sensible
default. We had changed to 1 to debug the production issues, but now
we can move back to this value. Note that this value should be same
as the CPU limit we have in k8s.

Reference:
https://opennmt.net/CTranslate2/python/ctranslate2.Translator.html

Add some test code with timing to debug this in future.

Change-Id: Ib76787a7f85e248d195df1641f69f7429bdd1ab8