ci: introduce ruff linting, black formatting, tox test wrapping
Use ruff for linting the source files. We did not had any kind
of linting so far. As the repository is getting active and more
source code is getting added it is important to have consistant
coding and formatting guidelines are enforced. black does the
formatting.

In CI, we need to run both linting and unit tests. Use tox as the
unified test interface for both of these. Blubber then just call
tox to execute both.

To get the CI passing some code changes are done in this patch.
But no changes related to functionality.

* Automatic code formatting, line length fixes by black
* pyttest update to use configtest for common text fixture
* remove unused wsgi.py
* add ruff, black, tox as dev dependencies

Change-Id: Ia14b25912cb3603334716b5e1a9cfaf9e74b23c5