html: rewrite the markup fixup system
The previous version of html markup fixup system was a very simple one.
It tried to find headers, paragragraphs and a predefined set of
annotations inside paragraphs. The inline annotations were annotated
after the translation using a fuzzy-matching approach.

This works for simple html files. But when this was used with cxserver
MT client, several problems were found. the paragraphs can have nested
nodes(like references). And when tried with realworld examples of webpages
several other shortcomings were found. This commit is a major rewrite of
the parsing logic.

Parsing:

* Make the traversal recursive so the deep nesting of nodes is solved
* Avoid the definition of block, inline tags, but infer this behavior
  by looking at the tag and contents. An <a> tag can be simple link, but
  it can also contain deeply nested html nodes.
* Improve the fuzzy matching logic to include levenshtein distance as
  measure of similarity. This greatly improves the matching logic for
  many languages. For example, the suffix inflection handling was not
  enough for languages like Spanish
* Restrict the markup fixup only when it is really needed. For example,
  handle the <ul><li> markup in simple way as <ul> is just wrapper and
  mixed nodes cannot appear inside <ul>
* Respect notranslate class to skip translation

Testing:

* Improve the "reverse" translator so that fullstops are not added
  unnecessarily.
* Add more tests and update tests as per update to "reverse" translator

Webpage translation:

* When the whole page is translated, apply html translation only to the
  <body> of page.

UI:

* Fix a bug that causes js error while doing webpage translation.

Change-Id: I8baa028bcb62aa738e64fc3f8e8a48f476021776