html: Avoid text extraction for non-translatable nodes
For example, when a sentence has a reference at end, the plain text
extraction of paragraph becomes "sentence[57]". This should not happen
as references are non-translatable.
To fix this, we subclass the Tag class of BS4, and skips non translatable
nodes from get_text() output.

Some test cases for this situation is also added. Previously the formatting
of HTML was affecting the output of translation. For example, new lines
before references was causing different output than when there is no new line.

Change-Id: I4dd5beea109c426cc8d4268a1365915588a13344