Improve reference handling in html adaptation
In a previous commit, while calculating plain text version of an html
node, ignorable(non-translatable) nodes were excluded. This resulted
text associated with references('[1]' etc) missing in plain text translation.
This caused default behavior of placing references at end of sentence.
This issue also became more visible in the case where an html paragraph is not
sentence segmented. All references went to the end of paragraph.

Fix this by not ignoring the plain text snippets of non-translatable elements.

However, this fix resurfaced the original issue - references appearing as
plain text values like [56] in translation. The root cause of this was
an inefficiency in fuzzy matching where punctuation like [ and ] are not
considered as search endings for a match. A few more issues noticed while
fixing: (a) case-sensitive ngram matching, (b) punctuation not ignored
in ngram matching. They are also fixed.

These issues were discovered as a result of POC done for T340956

Change-Id: I03790220e20721f87b03facf57ccb9363dcd44c4