server : rewrite using FastAPI
To take advantage of ASGI webserver, and thereby increase
the concurrency and performance, replace flask using fastapi.
This also makes the service in alignment with the Liftwing's
kserve ML inference architecture.

This is a drop-in replacement and no change in functionality
is expected.

In follow up patches, we will make use of fastapi's automatic
openapi spec generation and swagger UI. It is already in place
at /docs path, but need more data modeling to work.

Change-Id: I793cbd4119b5b9576c0628d913e545b8a1d36aa3