Improvements for rich text adaptation - references
References are often misplaced, missed, inserted in the translations
from NMT models:

1. indictrans2 models consistently remove all references from the
   source sentences
2. NLLB models rename the references. For example [8] will
   appear as [1] in the translation output

References are crucial, so this patch adds safeguards to ensure
nothing is missed nor added.

Approach:

At the pre-processing stage, collect all references in the source
sentence and store them. After translation, check if these references
are present in the translated sentence. If not, add them to the end of
the sentence.

Also, remove all references from the translated sentence that are not
present in the source sentence.

Tests:
For unit testing, we use a 'reversetrans' dummy model. We modify this
model so that it exhibit the two problems listed above regarding
references. It will add spurious references and it will remove
existing references. Even with this naughty reverse translator,
all tests should pass as all the models now guard against this behavior.

This patch can also be tested using the mint web interface 'html' translation
feature. Copy some content from wikipedia, with references. Paste it in
the rich text editor of html translation, observe the output.

Bug: T353791
Bug: T340956
Change-Id: I175bddc87eec0335e705033b1d8d8d2fe11a8dce