Improvements for rich text adaptation - repeated annotations
When inline annotations are repeated, it was observed that the markup is always restored on first occurrence.

For example, if input content has a link and that repeats at end
of content, annotation will be restored for the first item, leaving the last.

Same is the case of references when references are repeated in the input
content.

To fix this, keep track of restored annotation and avoid duplicate restorations.

Test included.

Bug: T353791
Bug: T340956
Change-Id: I4842fb3123261e3dc8082170bb298eb956070e31