html: Fix bug in repeated annotation translation
When annotations such as references are repeated in the source html,
the markup re-apply algorithm is skipping the second and subsequent
occurrences. This is because the algorithm is using the index of the
last fixup as the starting point for the next fixup. This indexes are
stored in a cache, but it was not cleared when we move to next siblings.

This caused wrong location for search start causing search misses.

This patch fixes the issue by clearing the cache when we move to next
sibling.

We introduced this algorithm in a previous commit 9274221d, but that
was only covering the situation where the annotation is repeated in same node.

Tests are added to cover the issue.

Bug: T363563
Change-Id: Ic887c634abc398fccb7b9d90d4d9da05f57540e6