html: Constrain the fuzzy matching to avoid broken html
In the case of IPA or languages like Igbo,  there are situations where the query string is a punctuation like single quote. When a fuzzy matching is done with the key, it can potentially match the json representation inside data-mw attributes of Mediawiki specific tags. And in such cases if a closing tag is introduced in between data-mw value, it will result a broken html as observed in T368521

Avoid this situation with the following constraints:
*  Make all mediawiki specific mw:transclusion nodes and nodes with data-mw non translatable
* In fuzzy matching check if key is a punctuation and avoid it from fuzzy matching

Bug: T368521
Change-Id: Ic4d71905c43c8d4d55e6ed013f267804f87b41da