Remove format from ecsformatter
Currently all logs are logged as NOTICE in Logstash even if they are
ERROR, INFO, WARN etc.

My assumption is that ecs_logging.StdlibFormatter is already designed
to automatically generate an ECS complaint log. Setting format to
%{message}s breaks this behavior.

If this works, we need to change it in the recommendation-api as well

Bug: T400562
Change-Id: I8c0531dc8a84cf6416647483a298c36ec49c9302